Clarify installation and data prep instructions
Adds an installation section to the README with the minimal packages
required for data preparation and the tidy data notebook.

Expands the initial datasets section to describe which CDC case counts
data to use (with a URL to download from the command line) and which
commands to run to sanitize and filter GISAID metadata.

Also, removes extraneous whitespace.